requirements.txt
python-dotenv==1.0.0
requests==2.32.4
urllib3==2.2.2
